moved the myuser.py methods into serializer file
