+++ accounts/serializers.py
@@ -1,5 +1,21 @@
 from rest_framework import serializers
 from django.contrib.auth.models import User
+from rest_framework import serializers
+from accounts.models import MyUser
+from app.models.section import Section
+from app.models.course_member import CourseMember
+
+class StudentMemberSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = MyUser
+        fields = ["id", "first_name", "last_name"]
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = MyUser
+        fields = "__all__"
+
 
 
 class UserSerializer(serializers.HyperlinkedModelSerializer):
